+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [0.11.0] - WIP
 ### Changed
 - Using new Matplotlib `pcolormesh` API introduced in v3.7.0 to plot sky map with projections faster with less memory
-- Add option to plot grid in sky map with projection (e.g., "aitoff")
+- Add option to plot galactic grid, equatorial grid and ecliptic grid in sky map
 
 ## [0.10.0] - 1 March 2023
 ### Changed